feat(deployer): split up deployer error enum
